refactor: split G-methods into _citation_methods.py, narrow warnings
- Extract 8 compute functions (G1-G8) from _divergence_citation.py into
  _citation_methods.py. The parent module keeps loading + graph infra.
  _divergence_citation.py: 555 -> 141 lines.
- Update dispatcher registry to point to _citation_methods.
- Remove blanket warnings.filterwarnings("ignore") from all three
  divergence modules. G5 uses catch_warnings() locally for curve_fit.
- Finding 6 (_incremental_graphs mutable graph): all 6 callers respect
  the contract (read within loop body, never store). G2/G8 call
  .to_undirected() which copies. No change needed.
