+++ scripts/_divergence_lexical.py
@@ -9,8 +9,6 @@ Methods:
 
 """
 
-import warnings
-
 import numpy as np
 import pandas as pd
 from pipeline_loaders import load_analysis_corpus
@@ -20,8 +18,6 @@ from sklearn.feature_extraction.text import TfidfVectorizer
 from utils import get_logger
 
 log = get_logger("_divergence_lexical")
-
-warnings.filterwarnings("ignore", category=FutureWarning)
 
 
 # ── Data loading ───────────────────────────────────────────────────────────
