+++ scripts/compute_divergence.py
@@ -46,56 +46,56 @@ METHODS = {
     "L2": ("_divergence_lexical", "compute_l2_novelty", "lexical", False, False),
     "L3": ("_divergence_lexical", "compute_l3_bursts", "lexical", False, False),
     "G1_pagerank": (
-        "_divergence_citation",
+        "_citation_methods",
         "compute_g1_pagerank",
         "citation",
         False,
         True,
     ),
     "G2_spectral": (
-        "_divergence_citation",
+        "_citation_methods",
         "compute_g2_spectral",
         "citation",
         False,
         True,
     ),
     "G3_coupling_age": (
-        "_divergence_citation",
+        "_citation_methods",
         "compute_g3_age_shift",
         "citation",
         False,
         True,
     ),
     "G4_cross_tradition": (
-        "_divergence_citation",
+        "_citation_methods",
         "compute_g4_cross_trad",
         "citation",
         False,
         True,
     ),
     "G5_pref_attachment": (
-        "_divergence_citation",
+        "_citation_methods",
         "compute_g5_pa_exponent",
         "citation",
         False,
         True,
     ),
     "G6_entropy": (
-        "_divergence_citation",
+        "_citation_methods",
         "compute_g6_entropy",
         "citation",
         False,
         True,
     ),
     "G7_disruption": (
-        "_divergence_citation",
+        "_citation_methods",
         "compute_g7_disruption",
         "citation",
         False,
         True,
     ),
     "G8_betweenness": (
-        "_divergence_citation",
+        "_citation_methods",
         "compute_g8_betweenness",
         "citation",
         False,
